Added 4 rooms and their exits
Added cafeteria which has an exit to its west of a kitchen and to its south a water closet to go along with the European theme. I added the water closet with an exit to its north of a cafeteria and to its west of a janitor closet. I added the janitor closet with an exit to the north of a kitchen and an exit to the east of a water closet. Lastly, I added the kitchen which has a exit to the east of a cafeteria and an exit to the south of a janitor closet.